refactor(client): one composeAvailableModels pipeline for every model picker
The model-availability chain existed three times: inline in
use-chat-session, in use-available-eval-models (org-wide query only, no
guest locks, no Ollama append), and in use-host-agent-models (added by
#2614) — the host copy kept in sync with chat by comment only.

Consolidate on the most correct semantics (the #2614 hook's):

- New chat-v2/shared/available-models.ts owns composeAvailableModels plus
  the moved applyGuestModelLocks / appendDetectedLocalOllamaModels /
  GUEST_LOCKED_MODEL_REASON. It lives next to (not inside) model-helpers
  so existing tests can keep mocking the per-source builders while the
  composition stays real.
- useDetectedOllamaModels extracts the 30s Ollama poll (with the
  hosted-mode bail and a cancellation guard the chat copy lacked).
- useAvailableModels({ projectId? }) replaces both useHostAgentModels and
  useAvailableEvalModels; the optional projectId preserves the eval
  surfaces' deliberate run/suite project scoping, defaulting to the
  active project.
- use-chat-session keeps its chatbox-aware org-config source but now
  feeds the shared composer.

Behavior change (evals only, intentional): eval suite/judge pickers now
use the project-scoped org config query with org-wide fallback (the
org-id-only query was the regression #2614's test pinned), apply guest
locks, append detected local Ollama models, and honor org config outside
hosted mode — matching chat and the client builder.
