+++ mcpjam-inspector/client/src/hooks/use-eval-tab-context.ts
@@ -1,7 +1,7 @@
 import { useMemo } from "react";
 import { useSharedAppState } from "@/state/app-state-context";
 import { useProjectMembers } from "@/hooks/useProjects";
-import { useAvailableEvalModels } from "@/hooks/use-available-eval-models";
+import { useAvailableModels } from "@/hooks/use-available-models";
 
 export function useEvalTabContext({
   isAuthenticated,
@@ -20,12 +20,14 @@ export function useEvalTabContext({
   const appState = useSharedAppState();
   // Scope to the requested project so model availability follows that project's
   // org rather than whatever happens to be the globally-active project.
+  // (`useAvailableModels` falls back to the active project when null.)
   const scopedProjectId = projectId ?? appState.activeProjectId ?? null;
   const scopedProject = scopedProjectId
     ? appState.projects?.[scopedProjectId]
     : undefined;
+  // Still returned to callers; the models hook re-derives it internally.
   const organizationId = scopedProject?.organizationId ?? null;
-  const { availableModels } = useAvailableEvalModels(organizationId);
+  const { availableModels } = useAvailableModels({ projectId });
   const { members, canManageMembers } = useProjectMembers({
     isAuthenticated,
     projectId,
